修改多规格，设置价格区间，修改订单

+++ ruoyi-modules/ruoyi-flower/src/main/java/org/dromara/flower/domain/FolwerProduct.java
@@ -60,12 +60,12 @@ public class FolwerProduct extends TenantEntity {
     /**
      * 原价
      */
-    private Long oriPrice;
+    private BigDecimal oriPrice;
 
     /**
      * 划线价
      */
-    private Long derlinePrice;
+    private BigDecimal derlinePrice;
 
     /**
      * 规格类型 默认是0，表示单规格，1表示多规格
