+++ ruoyi-modules/ruoyi-flower/src/main/java/org/dromara/flowerapplet/domain/FolwerAppletProduct.java
@@ -9,6 +9,7 @@ import org.dromara.common.translation.annotation.Translation;
 import org.dromara.common.translation.constant.TransConstant;
 
 import java.io.Serial;
+import java.math.BigDecimal;
 
 /**
  * 小程序端商品管理对象 folwer_product
@@ -58,12 +59,12 @@ public class FolwerAppletProduct extends TenantEntity {
     /**
      * 原价
      */
-    private Long oriPrice;
+    private BigDecimal oriPrice;
 
     /**
      * 划线价
      */
-    private Long derlinePrice;
+    private BigDecimal derlinePrice;
 
     /**
      * 规格类型 默认是0，表示单规格，1表示多规格
